Allow Switching Components Between Tab
