+++ src/components/search.component.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Search = ({ properties }:any) => {
+    return (
+        <div>Search Component</div>
+    )
+}
+
+export default Search
